+++ app/actions/index.jsx
@@ -2,4 +2,5 @@ export { default as app } from './app';
 export { default as db } from './db';
 export { default as blockDefs } from './blockDefs';
 export { default as menu } from './menu';
+export { default as message } from './message';
 export { default as user } from './user';