+++ app/components/App/App.jsx
@@ -6,8 +6,8 @@ import { BrowserRouter } from 'react-router-dom';
 import AppContext from '../AppContext';
 import ErrorHandler from '../ErrorHandler';
 import Main from '../Main';
+import Message from '../Message';
 import SideNavigation from '../SideNavigation';
-import TitleBar from '../TitleBar';
 
 /**
  * The main entry point of the React app.
@@ -22,9 +22,9 @@ export default class App extends React.Component {
           <MuiThemeProvider theme={createMuiTheme()}>
             <BrowserRouter basename={new URL(document.baseURI).pathname}>
               <AppContext>
-                <TitleBar />
                 <SideNavigation />
                 <Main />
+                <Message />
               </AppContext>
             </BrowserRouter>
           </MuiThemeProvider>