+++ app/components/Page/Page.jsx
@@ -1,9 +1,13 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import { FormattedMessage } from 'react-intl';
 
 import checkScope from '../../utils/checkScope';
 import Block from '../Block';
 import Login from '../Login';
+import TitleBar from '../TitleBar';
+import messages from './messages';
+import styles from './Page.css';
 
 /**
  * Render an app page definition.
@@ -52,17 +56,31 @@ export default class Page extends React.Component {
   };
 
   render() {
-    const { location, page, user } = this.props;
+    const { hasErrors, location, page, user } = this.props;
     const { dialog } = this.state;
 
     const { counter } = this;
 
     if (!checkScope(page.scope, user)) {
-      return <Login />;
+      return (
+        <React.Fragment>
+          <TitleBar>{page.name}</TitleBar>
+          <Login />
+        </React.Fragment>
+      );
+    }
+
+    if (hasErrors) {
+      return (
+        <p className={styles.error}>
+          <FormattedMessage {...messages.error} />
+        </p>
+      );
     }
 
     return (
       <React.Fragment>
+        <TitleBar>{page.name}</TitleBar>
         {page.blocks.map((block, index) => (
           <Block
             // As long as blocks are in a static list, using the index as a key should be fine.