+++ blocks/amsterdam-splash/index.jsx
@@ -25,20 +25,24 @@ attach(({ actions, data }) => {
   }, ANIMATION_LENGTH);
   const root = <div className={styles.root}>{loading}</div>;
   actions.load.dispatch(data).then(
-    () => {
-      root.replaceChild(
-        <header className={styles.content}>
-          <div className={styles.circle}>
-            <img alt="Success" className={styleMedia.icon} src={check} />
-          </div>
-          <h2 className={styles.header}>Gelukt</h2>
-          <span className={styles.subheader}>Dankjewel</span>
-        </header>,
-        loading,
-      );
+    response => {
       setTimeout(() => {
-        actions.success.dispatch(data);
-      }, 2e3);
+        root.replaceChild(
+          <header className={styles.content}>
+            <div className={styles.circle}>
+              <img alt="Success" className={styleMedia.icon} src={check} />
+            </div>
+            <h2 className={styles.header}>Gelukt</h2>
+            <span className={styles.subheader}>Dankjewel</span>
+          </header>,
+          loading,
+        );
+        root.classList.add(styles.done);
+
+        setTimeout(() => {
+          actions.success.dispatch(response);
+        }, 2e3);
+      }, 4e3);
     },
     () => {
       const button = (
@@ -50,7 +54,9 @@ attach(({ actions, data }) => {
         <header className={styles.content}>
           {button}
           <h2 className={styles.header}>Fout</h2>
-          <span className={styles.subheader}>Probeer het opnieuw</span>
+          <span className={styles.subheader}>
+            Druk op de bovenstaande knop om het opnieuw te proberen
+          </span>
         </header>
       );
       button.addEventListener('click', () => {