+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-11-02
+
 ### Added
 
 - API: Implement error reporting using Sentry.
+- Docker: Publish the appsemble/appsemble image on the public Docker Hub.
 - Editor: Add support for uploading app icons.
 - Frontend: Implement error reporting using Sentry.
+- Frontend: Add logout button in the side menu
+
+### Changed
+
+- Frontend: Show the page title instead of the app title.
 
 ## [0.1.0] - 2018-10-19